[Update] Year 기준으로 꾸준함에 따른 뱃지 조건 추가

+++ src/main/java/com/sparta/lafesta/badge/dto/BadgeRequestDto.java
@@ -18,5 +18,5 @@ public class BadgeRequestDto {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate conditionLastDay;
 
-    private Long conditionStandard;
+    private int conditionStandard;
 }
